LIBSAIL: Move sail_string_node alloc/destroy functions into string_node_private.h
