+++ src/libsail/sail_private.c
@@ -23,11 +23,6 @@
     SOFTWARE.
 */
 
-#include "config.h"
-
-#include <string.h>
-
-#include "sail-common.h"
 #include "sail.h"
 
 /*
@@ -163,62 +158,4 @@ sail_status_t allowed_write_output_pixel_format(const struct sail_write_features
 
     print_unsupported_write_pixel_format(pixel_format);
     SAIL_LOG_AND_RETURN(SAIL_ERROR_UNSUPPORTED_PIXEL_FORMAT);
-}
-
-sail_status_t alloc_string_node(struct sail_string_node **string_node) {
-
-    SAIL_CHECK_STRING_NODE_PTR(string_node);
-
-    void *ptr;
-    SAIL_TRY(sail_malloc(sizeof(struct sail_string_node), &ptr));
-    *string_node = ptr;
-
-    (*string_node)->value = NULL;
-    (*string_node)->next  = NULL;
-
-    return SAIL_OK;
-}
-
-void destroy_string_node(struct sail_string_node *string_node) {
-
-    if (string_node == NULL) {
-        return;
-    }
-
-    sail_free(string_node->value);
-    sail_free(string_node);
-}
-
-void destroy_string_node_chain(struct sail_string_node *string_node) {
-
-    while (string_node != NULL) {
-        struct sail_string_node *string_node_next = string_node->next;
-
-        destroy_string_node(string_node);
-
-        string_node = string_node_next;
-    }
-}
-
-sail_status_t split_into_string_node_chain(const char *value, struct sail_string_node **target_string_node) {
-
-    struct sail_string_node **last_string_node = target_string_node;
-
-    while (*(value += strspn(value, ";")) != '\0') {
-        size_t length = strcspn(value, ";");
-
-        struct sail_string_node *string_node;
-
-        SAIL_TRY(alloc_string_node(&string_node));
-
-        SAIL_TRY_OR_CLEANUP(sail_strdup_length(value, length, &string_node->value),
-                            /* cleanup */ destroy_string_node(string_node));
-
-        *last_string_node = string_node;
-        last_string_node = &string_node->next;
-
-        value += length;
-    }
-
-    return SAIL_OK;
 }
